pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hireme"
version = "0.1.0"
description = "Autonomous internship/co-op application agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LLM and orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openrouter>=0.1.0",

    # Vector store
    "chromadb>=0.5.0",

    # Browser automation
    "playwright>=1.45.0",

    # PDF generation
    "jinja2>=3.1.0",
    "weasyprint>=62.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",

    # Type hints
    "typing-extensions>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[project.scripts]
hireme = "hiremeAI.graph:main"
hireme-schedule = "hiremeAI.scheduler:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hiremeAI*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
